+++ src/main/java/com/example/mssaem_backend/global/common/Comment.java
@@ -15,7 +15,7 @@ public interface Comment {
     Discussion getDiscussion();
     Board getBoard();
 
-    void setParentComment(Integer commentId);
+    void setParentComment(Long commentId);
 
     void deleteComment();
 }
